Fix font in button causing segmentation fault
Pushing of button to vector caused the path of the font to change.
This was fixed with referencing non-changing shared pointer.
This could also probably have been fixed by just not using
push_back but using of pointer (or reference) is ideal due textures,
fonts, and audio take a lot of space and you do not want to have
them everywhere.